Move payment_intent to Helper
