+++ spec/models/spree/payment_spec.rb
@@ -1128,17 +1128,4 @@ describe Spree::Payment do
       expect(payment.captured_at).to be_present
     end
   end
-
-  private
-
-  def payment_intent(amount, status)
-    JSON.generate(
-      object: "payment_intent",
-      amount:,
-      status:,
-      charges: { data: [{ id: "ch_1234", amount: }] },
-      id: "12345",
-      livemode: false
-    )
-  end
 end
